[alphas] ** COMPLETE ALPHAS ** further progress in backporting heft - does not build yet, see #439
I will merge this anyway as standalone cudacpp for SM physics works fine
(there is one exception, uudd fails, see #440 - I will fix that a posteriori).

Note also that alphas from madevent are still not integrated, and so ggtt.mad fails to build for instance, #441.
This will be the next big thing.