[gqttq] ** COMPLETE GQTTQ** revert tmad test and log for gqttq from P1_gux_ttxux to P1_gu_ttxu
The gqttq tests fail anyway and will need to be fixed (#630).
However, this completes the addition of gq_ttq as a new process to the repo.
In particular it includes proof that Olivier's "split_nonidentical_grouping" #619 fixes the gqttq builds.
It also includes a lot of cleanup for "nprocesses" (#272 and #343)

Revert "[gqttq] retry the tmad gqttq test with the P1_gu_ttxu directory - the test continues to fail (#630)"
This reverts commit 2dea1f76a8d9faa5fbc20bc8967c27da5b1f1762.

Revert "[gqttq] temporarely use P1_gu_ttxu instead of P1_gux_ttxux for gqttq tmad tests"
This reverts commit 399d5d8ded3ada6fe5bab69cfce755d7dfc49ef5.