Add a script for running code generator from mg5 interface.
In order to facilitate the code generation, add a script to generate the
processes in this repo. Pass all processes to be generated as argument,
e.g:
  ./runCodegen.sh gg_tt.mad ee_mumu.mad

It is assumed that each process directory contains mg5.in with all
madgraph commands that are needed to generate the corresponding process.

The script will try to stage the changes, so they can be committed.