[jtmk] in CODEGEN, partial fix of merge conflicts in cudacpp.mk (handling of CUDA_HOME and HIP_HOME)
Start enforcing and documenting NEW LOGIC FOR ENABLING AND DISABLING CUDA OR HIP BUILDS (AV 02.02.2024)
- As in the past, we first check for cudacc and hipcc in CUDA_HOME and HIP_HOME; and if CUDA_HOME or
HIP_HOME are not set, we try to determine them from the path to cudacc and hipcc. **FIXME** : in the
future, it may be better to completely remove the dependency of this makefile on CUDA_HOME and HIP_HOME,
and only rely on whether nvcc and hipcc are in PATH.
- In the old implementation, by default the C++ targets for one specific AVX were always built together
with either CUDA or HIP. If both CUDA and HIP were installed, then CUDA took precedence over HIP, and the
only way to force HIP builds was to disable CUDA builds by setting CUDA_HOME to an invalid value. Similarly,
C++-only builds could be forced by setting CUDA_HOME and/or HIP_HOME to invalid values. A check for an
invalid nvcc in CUDA_HOME or an invalid hipcc HIP_HOME was necessary to ensure this logic, and had to be
performed _before_ choosing whether C++/CUDA, C++/HIP or C++-only builds had to be executed.
- In the new implementation (PR #798), separate individual builds are performed for one specific C++/AVX
mode, for CUDA or for HIP. The choice of the type of build is taken depending on the value of the BACKEND
variable (replacing the AVX variable). For the moment, it is still possible to override the PATH to nvcc
and hipcc by externally setting a different value to CUDA_HOME and HIP_HOME (although this will probably
disappear as mentioned above): note also that the check whether nvcc or hipcc exist still needs to be
performed _before_ BACKEND-specific configurations, because CUDA/HIP installations also affect C++-only
builds, for NVTX and curand-host in the case of CUDA, and for rocrand-host in the case of HIP.
- Note also that the REQUIRE_CUDA variable (#443) is now (PR #798) no longer necessary, as it is now
equivalent to BACKEND=cuda. Similarly, there is no need to introduce a REQUIRE_HIP variable.