Add a makefile wrapper to avoid double -f while doing make
Makefile wrapper used in case of madevent to join together both makefile_orig
and cudacpp_overlay.mk. This is to avoid doing make -f makefile -f cudacpp.mk,
keeping the possibility to just do `make`:
- makefile (after code generation by MadGraph) is renamed to makefile_original.mk
- makefile_wrapper.mk (copied from cudacpp) is symlinked as makefile
The logic is: in each SubProcesses/P* directory we have
- makefile -> ../makefile (this is the original makefile)
- cudacpp_overlay.mk -> ../cudacpp_overlay.mk
While in the SubProcesses directory we had to:
- makefile (this is the original makefile)
- cudacpp_overlay.mk
- makefile_wrapper.mk
So, in each SubProcesses/P* we had to (done in model_handling.py) symlink from
SubProcesses/P*/makefile_original.mk to SubProcesses/makefile_original.mk
(it will be broken initially).
While in SubProcesses we had to (done in output.py):
- copy makefile to makefile_original.mk (this fixes the previous symlink)
- remove makefile
- symlink from makefile_wrapper.mk to makefile
So that now the situation in each SubProcesses/P* is:
- makefile -> ../makefile (this is the makefile_wrapper.mk)
- cudacpp_overlay.mk -> ../cudacpp_overlay.mk
- makefile_original.mk -> ../makefile_original.mk (this is the original makefile)